# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hospital API",
        "description": "API documentation for the Hospital application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Default Server"
        }
    ],
    "paths": {
...
        "/api/patients": {
            "get": {
                "tags": [
                    "Patients"
                ],
                "summary": "List patients",
                "operationId": "0df0d65f049161931654c3e188ca8f93",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Patient"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Patients"
                ],
                "summary": "Create patient",
                "operationId": "2149729f1bf1af8877352f04be3a683f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePatientRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Patient"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/patients/{id}": {
            "get": {
                "tags": [
                    "Patients"
                ],
                "summary": "Get patient by ID",
                "operationId": "900a000f9712c06fdf96fa64333bbaff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Patient"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Patients"
                ],
                "summary": "Update patient",
                "operationId": "49d64723ad7c62df6eddc6fa8a778755",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePatientRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Patient"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Patients"
                ],
                "summary": "Delete patient",
                "operationId": "0187379ab8f57fd78eeb510643be993b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/payments": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "List payments",
                "operationId": "afcbc19a07bf33c2fcb73b5c438b476c",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Payment"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Create payment",
                "operationId": "985b848b4cb6ca6115cf4ea71e314c0c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePaymentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Payment"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/payments/{id}": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Get payment by ID",
                "operationId": "33da95d13a42b4c3898427d7cb9516a8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Payment"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Payments"
                ],
                "summary": "Update payment",
                "operationId": "4e85438414b16b3c35ad3776c5ccc11f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePaymentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Payment"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Payments"
                ],
                "summary": "Delete payment",
                "operationId": "4db359e13e73e2832ead474820546c40",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StorePatientRequest": {
                "required": [
                    "firstname",
                    "lastname",
                    "email",
                    "password"
                ],
                "properties": {
                    "firstname": {
                        "type": "string"
                    },
                    "lastname": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "minimum": 8
                    },
                    "phone": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "occupation": {
                        "type": "string"
                    },
                    "image": {
                        "type": "string"
                    },
                    "gender": {
                        "type": "string",
                        "enum": [
                            "Male",
                            "Female",
                            "Other"
                        ]
                    },
                    "city": {
                        "type": "string"
                    },
                    "state": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive"
                        ]
                    },
                    "blood_group": {
                        "type": "string",
                        "enum": [
                            "A+",
                            "A-",
                            "B+",
                            "B-",
                            "AB+",
                            "AB-",
                            "O+",
                            "O-"
                        ]
                    },
                    "date_of_birth": {
                        "type": "string",
                        "format": "date"
                    },
                    "age": {
                        "type": "integer",
                        "minimum": 0
                    },
                    "height": {
                        "type": "integer"
                    },
                    "weight": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "StorePaymentRequest": {
                "required": [
                    "paymentable_type",
                    "paymentable_id",
                    "amount"
                ],
                "properties": {
                    "paymentable_type": {
                        "type": "string"
                    },
                    "paymentable_id": {
                        "type": "integer"
                    },
                    "payment_method": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "integer",
                        "minimum": 0
                    },
                    "currency": {
                        "type": "string"
                    },
                    "payment_status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "paid",
                            "failed"
                        ]
                    },
                    "transaction_id": {
                        "type": "string"
                    },
                    "payment_date": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "meta": {
                        "type": "array",
                        "items": {}
                    }
                },
                "type": "object"
            },
...
            "UpdatePatientRequest": {
                "properties": {
                    "firstname": {
                        "type": "string"
                    },
                    "lastname": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "minimum": 8
                    },
                    "phone": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "occupation": {
                        "type": "string"
                    },
                    "image": {
                        "type": "string"
                    },
                    "gender": {
                        "type": "string",
                        "enum": [
                            "Male",
                            "Female",
                            "Other"
                        ]
                    },
                    "city": {
                        "type": "string"
                    },
                    "state": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive"
                        ]
                    },
                    "blood_group": {
                        "type": "string",
                        "enum": [
                            "A+",
                            "A-",
                            "B+",
                            "B-",
                            "AB+",
                            "AB-",
                            "O+",
                            "O-"
                        ]
                    },
                    "date_of_birth": {
                        "type": "string",
                        "format": "date"
                    },
                    "age": {
                        "type": "integer",
                        "minimum": 0
                    },
                    "height": {
                        "type": "integer"
                    },
                    "weight": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "UpdatePaymentRequest": {
                "properties": {
                    "paymentable_type": {
                        "type": "string"
                    },
                    "paymentable_id": {
                        "type": "integer"
                    },
                    "payment_method": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "integer",
                        "minimum": 0
                    },
                    "currency": {
                        "type": "string"
                    },
                    "payment_status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "paid",
                            "failed"
                        ]
                    },
                    "transaction_id": {
                        "type": "string"
                    },
                    "payment_date": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "meta": {
                        "type": "array",
                        "items": {}
                    }
                },
                "type": "object"
            },
...
            "Patient": {
                "title": "Patient",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "firstname": {
                        "type": "string",
                        "example": "John"
                    },
                    "lastname": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "1234567890"
                    },
                    "address": {
                        "type": "string",
                        "example": "123 Main St"
                    },
                    "occupation": {
                        "type": "string",
                        "example": "Doctor"
                    },
                    "image": {
                        "type": "string",
                        "example": "https://example.com/image.jpg"
                    },
                    "gender": {
                        "type": "string",
                        "example": "Male"
                    },
                    "city": {
                        "type": "string",
                        "example": "New York"
                    },
                    "state": {
                        "type": "string",
                        "example": "New York"
                    },
                    "country": {
                        "type": "string",
                        "example": "USA"
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    },
                    "blood_group": {
                        "type": "string",
                        "example": "A+"
                    },
                    "date_of_birth": {
                        "type": "string",
                        "format": "date",
                        "example": "1990-01-01"
                    },
                    "age": {
                        "type": "integer",
                        "example": "30"
                    },
                    "height": {
                        "type": "integer",
                        "example": "180"
                    },
                    "weight": {
                        "type": "integer",
                        "example": "70"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            },
            "Payment": {
                "title": "Payment",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "paymentable_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "paymentable_type": {
                        "type": "string",
                        "example": "App\\\\Models\\\\User"
                    },
                    "payment_method": {
                        "type": "string",
                        "example": "credit_card"
                    },
                    "amount": {
                        "type": "integer",
                        "example": "100"
                    },
                    "currency": {
                        "type": "string",
                        "example": "USD"
                    },
                    "payment_status": {
                        "type": "string",
                        "example": "pending"
                    },
                    "transaction_id": {
                        "type": "string",
                        "example": "1234567890"
                    },
                    "payment_date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-01-01 00:00:00"
                    },
                    "meta": {
                        "type": "array",
                        "items": {},
                        "example": "[]"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}